Fix for MSVC SmallerTypeCheck runtime error
The assignment of carry to p in eval_multiply generates a runtime error in non-optimized MSVC builds which have SmallerTypeCheck enabled. As carry is twice the size of p and the top bits are non-zero the check errors saying that some bits are being ignored. This is Microsoft's recommended fix of masking the unwanted bits when assigning to the smaller type.