Fix detection of constructibility from types with protected dtor in C++03.
The declval function returns T by value in C++03, which requires the type
to have a public destructor in the context of is_explicitly_convertible_imp.
If this is not true the compilation hard-fails on some compilers (e.g. all
clang versions up to 7.0). The problem appears, for instance, when
boost::multiprecision::number is wrapped in boost::optional, the latter
derives from optional_base, which has a protected destructor. boost::optional
copy constructor needs to pass a reference to optional_base to optional_base
constructor, which at overload resolution tests multiprecision::number
conversion constructor and fails.

We fix the problem by generating a const reference by declval in C++03,
so the destructor is never tested.

Also, added missing includes and consistently use Boost.TypeTraits instead
of Boost.MPL througout the file.