Use const_iterator to obtain container value_type in is_byte_container.
The value_type typedef is quite widespread, even in types that are not ranges.
For example, Boost.Optional uses this typedef internally as a protected typedef,
which fails the compilation of is_byte_container.

We fix this by testing for the const_iterator typedef first. Then we obtain
the value type from the iterator instead of the container because we will be
using iterators to access the container elements anyway.