Fix Apple clang version detection false positives.
In number.hpp, a workaround is presented for old versions of Apple clang. Apple uses non-standard version numbers, which makes it difficult to rely on __clang_major__. The code first checked __APPLE_CC__ in an attempt to check "is this compiler using Apple's version numbering", but open source upstream clang versions also define this variable when compiling to Apple targets; as this workaround breaks some client code, this makes some codebases impossible to compile for Apple targets with upstream clang. The correct variable to check is __apple_build_version__.