Only set the precision if the precision has changed.
This reduces contention with multiple threads.  Previously, all
threads would unconditionally try to set the precision independently.
For the GMP backend, the global precision seems to be protected by a
mutex, so that would end up making all of the threads block when
setting precision.  This is true even if all of the threads are using
the same precision.  This is particularly problematic because it
happens for each and every arithmetic operation (e.g. adding two
numbers).

This does not fix the overall threading problem that a computation on
one thread can change the precision on another thread.  Or that after
one thread has reset the precision, another thread can undo that
reset.