Adjust CircleCI setup to account for Boost.Inspect build change.
In addition to changing how to build the inspect tool this change Also simplifies the build and test to avoid future config problems as it can more easily adjust to future CI changes. It bumps the docker image to what the current stable image is to auto-adjust to recent stable gcc compiler.